% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jokes_function.R
\name{yo_momma_api}
\alias{yo_momma_api}
\title{A function that returns a your-momma joke}
\arguments{
\item{yo_momma_api}{returns a random your-momma joke}
}
\value{
returns a random your momma joke from api.yomoma.info
yo_momma()
}
\description{
This function allows you to express your love of cats.
}
\keyword{jokes}
